Add multiplication operation in Oasis library
A "Multiply" operation has been implemented as part of the Oasis library. This included adding header and source files, creating multiplication functions in these new files, and including the "Multiply" module in the CMakeLists. Corresponding tests have been put in place in the main test script to validate the functionality.